backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
psycopg2-binary==2.9.10
sqlparse==0.5.3
djangorestframework>=3.12,<4.0
gunicorn>=20.1,<21.0
django-environ>=0.8,<1.0
beautifulsoup4==4.13.4
pandas==2.3.0
pytz==2024.2
Requests==2.32.4
python-dotenv>=1.0.0
django-cors-headers>=4.6.0
djangorestframework
pytest>=7.3.1,<8.0
pytest-django>=4.5.2,<5.0
